fix overflow in random by using longs instead of ints

+++ include/random.h
@@ -4,12 +4,12 @@
 
 bool rand_weighted_bool(double percent);
 bool rand_bool();
-double rand_seed(int seed);
+double rand_seed(long seed);
 double rand_seed_rand();
-int rand_int();
+long rand_int();
 double rand_double();
 double rand_double_range(double min, double max);
-int rand_int_range(int min, int max);
+long rand_int_range(long min, long max);
 double rand_power(double min, double max, double power);
 double rand_gauss(double min, double max, int gauss);
 #endif
